+++ sphinx/source/conf.py
@@ -4,8 +4,8 @@ import os
 project = 'Indexor'
 copyright = '2026, Aarav Agarwal'
 author = 'Aarav Agarwal'
-version = '0.5.1'
-release = '0.5.1'
+version = '0.5.2'
+release = '0.5.2'
 
 extensions = [
     "sphinx.ext.autodoc",